Allow for multiple ID generators to be run per email
If needed, this allows custom multi-id generation for emails by
specifying more than one generator (with a space between each).
This allows for seamless switching between short and long links.
The first generator will be used for the document ID, and all
generated IDs will be present in the 'permalinks' array.