re-align with old pony for cluster generator and unit tests
The general idea here is, if we find an email without a charset at all,
and we detect non-ascii characters in it, we assume it must be UTF-8 and
grab the raw bytes. We also convert it internally to a string for the
Body class, but we don't set the Body class' character set to anything.
This way, we keep the cluster generator happy by passing it bytes, while
keeping the rest happy by having a string representation that can be
unflowed. As DKIM does not use the msgbody itself, it won't be affected
by this change.