Faster code to clean up quotes
Right now the `*` in the regex to clean up quotes makes it rather slow for
large quoted texts. With this change it is slightly less 'precise' but
much faster, which might be a reasonable trade-off in this case?

You can reproduce at https://regexr.com/ by adding 50 or so newlines to
the example input.

The effect in Pony Mail is that such emails cannot be opened at all,
but produce some javascript error.