Add an option useful for personal email archiving. Handling List-Id's.  Ponymail likes to sort mails by the List-Id which makes a lot of sense where you have the thousands of Apache lists.  But with personal email, and certainly when you subscribe to various newsletters, or get bills, or spam that got into the archives then you end up with lots of list id's that are only used once or twice or are not useful.  Working on open source projects there's lots of lists that I'm on that I want the email to get archived, but it would be nice if it was separated out in the Ponymail UI.  So really I needed the ability to have an 'allow list' of list id's that I want to have separate, with everything else defaulting to a generic list id (being my email address where all those mails came into).
If you run with an allow: list and run import-mbox with a --lid default then all list-ids will be replaced by your --lid default unless they match the text substring in the allow list.  All other list ids are displayed (so you can do a dry run and spot which you'd like on the allow list).  Any on the ignore list are not printed, nothing else is done with them, it's just so you can get a smaller list of unknown list-ids displayed to decide which to add to your allow list.