tools/import-mbox.py: un-break get_file() for Maildir
In most implementations of the Pyhon mailbox module, get_file()
only takes a single argument (two when counting the object's "self"
reference). Only the mbox implementation differs - it supports an
additional key-value argument "from_". Since we want to set this
argument we need to check which implementation is actually used.

This un-breaks support for Maildir imports. Previously, Maildir imports
would stumble over the additional argument and abort the import.

See
https://docs.python.org/3/library/mailbox.html#mailbox.Maildir.get_file
and
https://github.com/python/cpython/blob/main/Lib/mailbox.py
for more information.

Signed-off-by: Thilo Fromm <thilo.alexander@gmail.com>